backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-MySQLdb
cryptography